Add outline for page UI
