+++ src/Components/App/App.js
@@ -1,5 +1,8 @@
 import React, { Component } from 'react';
 import LandingPage from '../LandingPage/LandingPage';
+import ScrollSection from '../ScrollSection/ScrollSection';
+import Header from '../Header/Header';
+import CardContainer from '../CardContainer/CardContainer';
 import './App.css';
 
 class App extends Component {
@@ -14,7 +17,13 @@ class App extends Component {
   render() {
     return (
       <div className="App">
-        <LandingPage />
+        <main className="component-container">
+        <ScrollSection className="scroll-section"/>
+        <section className="right-section">
+        <Header className="header" />
+        <CardContainer className="card-container" />
+        </section>
+        </main>
       </div>
     );
   }
