+++ src/Components/CardContainer/card-container.css
@@ -1,0 +1,3 @@
+.card-container{
+  background-color: purple;
+}
