+++ src/Components/Header/Header.js
@@ -1,9 +1,10 @@
 import React from 'react';
+import './header.css'
 
 const Header = () => {
 
   return (
-    <div>
+    <div className="header-div">
     The Header
     </div>
   );
