+++ src/Components/ScrollSection/scroll-section.css
@@ -1,0 +1,4 @@
+.swapi-text-div{
+  background-color: blue;
+  height: 100vh;
+}
