--- src/Components/SideNav/SideNav.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const SideNav = () => {
-
-  return (
-    <div>
-    The SideNav
-    </div>
-  );
-}
-
-export default SideNav;
